extract MapRef and SetRef to a single class since they differ only in name, #278
